requirements.txt
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.7
Flask==3.0.0
google-api-core==2.28.1
google-api-python-client==2.186.0
google-auth==2.41.1
google-auth-httplib2==0.2.1
google-auth-oauthlib==1.2.3
googleapis-common-protos==1.71.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httplib2==0.31.0
idna==3.11
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==6.0.2
MarkupSafe==3.0.3
oauthlib==3.3.1
outcome==1.3.0.post0
packaging==25.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyee==13.0.0
pyparsing==3.2.5
PySocks==1.7.1
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.0.1
wsproto==1.2.0
